pyproject.toml
[build-system]
requires = [
    "setuptools"
]
build-backend = "setuptools.build_meta"

[project]
name = "ezrpg"
version = "2022.3.6.2"
description = "Make it slightly easier to run RPG games with Python"
readme = "README.rst"
authors = [{name = "Moshe Zadka", email = "moshez@zadka.club"}]
dependencies=["attrs", "toml"]

[project.optional-dependencies]
tests = ["virtue", "pyhamcrest", "coverage"]
mypy = ["mypy"]
lint = ["flake8", "black"]
docs = ["sphinx", "nbsphinx", "ipython"]

[project.license]
text = """
Permission is hereby granted, free of charge, to any person obtaining a
copy of this software and associated documentation files (the "Software"),
to deal in the Software without restriction, including without limitation the
rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
copies of the Software, and to permit persons to whom the Software is furnished
to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all
copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF
CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE
OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
"""

[project.urls]
Homepage = "https://github.com/moshez/ezrpg"
